# Leverage Block#mayHaveNull() in HashSemiJoinOperator