@aweisberg 2021-07-23
So one other thing that was brought up was the pre-computed [hash block optimization](https://github.com/trinodb/trino/blob/4b90f8e5a96b4e3b13a6289177967def25e97c82/core/trino-main/src/main/java/io/trino/operator/HashSemiJoinOperator.java#L188,L220) which would bring us to four different permutations of this loop.