@aweisberg 2021-07-23
@pettyjamesm can you update the commit message body of the last commit to call out that is also unwrapping the `LazyBlock` as an improvement?