@pettyjamesm 2021-07-23
Updated the comment, also added the SQL benchmark results which- while noisy and likely to underestimate the performance difference because of benchmarking environment specifics like the block types being monomorphic in an unrealistic manner still show that:
1. There's still a performance gain compared to baseline
2. C2 loop unswitching does appear to have occurred since performance is nearly identical to the manually unswitched version.